+++ built/grammar/AbstractRuntimeProgram.d.ts
@@ -1,13 +1,24 @@
 import TreeNode from "../base/TreeNode";
 import AbstractRuntimeNode from "./AbstractRuntimeNode";
 import types from "../types";
+import GrammarProgram from "./GrammarProgram";
 declare abstract class AbstractRuntimeProgram extends AbstractRuntimeNode {
     getProgramErrorsIterator(): IterableIterator<any>;
     getProgramErrors(): types.ParseError[];
     getInvalidKeywords(level?: any): string[];
+    _getAllSuggestions(): string;
+    getAutocompleteWordsAt(lineIndex: types.positiveInt, charIndex: types.positiveInt): {
+        startCharIndex: number;
+        endCharIndex: number;
+        word: string;
+        matches: {
+            text: string;
+            displayText: string;
+        }[];
+    };
     getProgramErrorMessages(): string[];
     getKeywordMap(): any;
-    getDefinition(): any;
+    getDefinition(): GrammarProgram;
     getKeywordUsage(filepath?: string): TreeNode;
     getInPlaceSyntaxTree(): string;
     getInPlaceHighlightScopeTree(): string;