+++ built/grammar/TreeNotationCodeMirrorMode.js
@@ -99,7 +99,8 @@ class TreeNotationCodeMirrorMode {
         return __awaiter(this, void 0, void 0, function* () {
             const cursor = cmInstance.getCursor();
             const codeMirrorLib = this._getCodeMirrorLib();
-            const result = yield this.autocomplete(cmInstance.getLine(cursor.line), cursor.line, cursor.ch);
+            const result = yield this._getParsedProgram().getAutocompleteWordsAt(cursor.line, cursor.ch);
+            console.log(result);
             return result.matches.length
                 ? {
                     list: result.matches,
@@ -107,44 +108,6 @@ class TreeNotationCodeMirrorMode {
                     to: codeMirrorLib.Pos(cursor.line, result.endCharIndex)
                 }
                 : null;
-        });
-    }
-    // todo: why is this async?
-    autocomplete(line, lineIndex, charIndex) {
-        return __awaiter(this, void 0, void 0, function* () {
-            const mode = this;
-            let start = charIndex;
-            let end = charIndex;
-            while (start && /[^\s]/.test(line.charAt(start - 1)))
-                --start;
-            while (end < line.length && /[^\s]/.test(line.charAt(end)))
-                ++end;
-            const input = line.slice(start, end).toLowerCase();
-            // For now: we only autocomplete if its the first word on the line
-            if (start > 0 && line.slice(0, start).match(/[a-z]/i))
-                return {
-                    startCharIndex: start,
-                    endCharIndex: end,
-                    matches: []
-                };
-            const program = mode._getParsedProgram();
-            const isChildNode = start > 0 && lineIndex > 0;
-            const nodeInScope = isChildNode ? program.getTopDownArray()[lineIndex].getParent() : program;
-            const grammarNode = nodeInScope.getDefinition();
-            // todo: add more tests
-            // todo: second param this.childrenToString()
-            // todo: change to getAutocomplete definitions
-            const matches = grammarNode.getAutocompleteWords(input).map(str => {
-                return {
-                    text: str,
-                    displayText: str
-                };
-            });
-            return {
-                startCharIndex: start,
-                endCharIndex: end,
-                matches: matches
-            };
         });
     }
     register() {