+++ releaseNotes.md
@@ -1,7 +1,13 @@
-19.3.1 / 2019-04-21
+19.3.0 / 2019-04-21
 ===================
+- New: new base methods nodeAtLine, getNodeInScopeAtCharIndex, getWordIndexAtCharacterIndex, getWordProperties, getWordBoundaryIndices,   getAllWordBoundaryCoordinates
+- New: on runtime programs: getAutocompleteWordsAt and getAllSuggestions
 - Fix: error highlight scope fixes
 - Breaking: getAutocompleteWords now takes an optional 2nd column index parameter
+- Breaking: instead of getAutocompleteWords use getAutocompleteWordsAt
+
+
+
 
 19.2.1 / 2019-04-20
 ===================